CryptContext: added "harden_verify" support
v1.6 deprecated the "min_verify_time" feature, since it required admins
to awkwardly hard-code the amount of time they wanted .verify() to take.
This was a useful feature, though, since it provided the ability for
contexts to mask which accounts had older, weaker, hashes.

This commit adds in a replacement feature: CryptContext now supports
a "harden_verify" flag, which causes it to autocalculate how long
the default hash should take, and make sure to spend at least that much time.

Another improvement is that this autocalc is performed at the context level,
not per hash like min_verify_time was, making it overall more reliable.
It also offers a new "context.dummy_verify()" call for faking the delay
for missing accounts.

Changes:

* Added "harden_verify" option to CryptContext,
  along with code to autocalculate min_verify_time delay via timing
  when context is first accessed.

  XXX: this timing code needs some battle-testing, and may be adjusted later.

* Added .dummy_verify() method, integrated it into .verify()
  and .verify_and_update().

* updated warnings & docs to report harden_verify as replacement
  for "min_verify_time" option.

* to help workflow for adding delay on missing accounts,
  .verify() and .verify_and_update() now support passing in hash=None,
  and treat it as "never verifying"